api/requirements.txt
azure-functions
requests
fastapi
uvicorn
